requirements.txt
discord.py[voice]
discord.py
youtube-dl
